# Use .count, not .length, for better performance