--- a/app/services/ems_dashboard_service.rb
+++ b/app/services/ems_dashboard_service.rb
@@ -51,7 +51,7 @@ class EmsDashboardService < DashboardService
         :id           => attr_hsh[attr] + '_' + @ems_id,
         :iconClass    => attr_icon[attr],
         :title        => attr_hsh[attr],
-        :count        => @ems.send(attr).length,
+        :count        => @ems.send(attr).count,
         :href         => get_url(ems_type, @ems_id, attr_url[attr]),
         :notification => {
           :iconClass => 'pficon pficon-error-circle-o',